# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/generate:
    post:
      operationId: generateAiContent
      tags: [ai]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AiGenerationInput"
      responses:
        "200":
          description: Generated educational content
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiGenerationResult"
        "402":
          description: Active subscription required
components:
...
  schemas:
...
    AiGenerationInput:
      type: object
      required: [type, subject, grade]
      properties:
        type:
          type: string
          enum: [lesson, homework, exam, qiyas]
        subject: { type: string }
        grade: { type: string }
        topic: { type: string }
        imageData:
          type: string
          description: Optional transient image data URL for solving a photographed homework question
          maxLength: 7000000
        fallbackTestNumber:
          type: integer
          minimum: 1
          maximum: 20
          description: Selects one of the 20 local fallback tests when the AI provider is unavailable
    AiGenerationResult:
      type: object
      required: [title, body, items]
      properties:
        title: { type: string }
        body: { type: string }
        items:
          type: array
          items:
            type: object
            required: [question, answer, explanation]
            properties:
              question: { type: string }
              answer: { type: string }
              explanation: { type: string }
              options:
                type: array
                description: Multiple-choice options for a generated practice question
                items: { type: string }
              skill: { type: string }
              difficulty:
                type: string
                enum: [سهل, متوسط, صعب]
        source:
          type: string
          enum: [ai, fallback]
        testNumber: { type: integer }
        testsAvailable: { type: integer }
        questionsPerTest: { type: integer }